Add simple items storage
